+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/listeners/VcsMappingChangedListener.kt
@@ -28,8 +28,8 @@ class VcsMappingChangedListener(val project: Project) : VcsRepositoryMappingList
                 .map { it.toString() }
                 .toSet()
 
-        AppSettings.instance.remotes.values
-            .mapNotNull { it.gitlabUrl }
+        AppSettings.instance.remotes
+            .mapNotNull { it.value.gitlabUrl }
             .forEach { projectSettings.gitlabUrls += it }
 
         if (projectSettings.gitlabUrl != null && projectSettings.gitlabUrls.contains(projectSettings.gitlabUrl)) {