+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Refactor code for getting remote URL and Gitlab API URL in `ResolveContext`.
 - Show notification about detected Gitlab API URL only once per project.
 - Do not show detected Gitlab API URL in the notification, because there can be multiple Gitlab API URLs in one project.
+- Show merged preview even if there are errors, if the response contains the merged yaml.
 
 ### Fixed
 - NPE when opening a folder as project.