+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+
+## [1.12.0] - 2023-05-15
+
 ### Added
 - Option to ignore specific error for a file.
 - Explanation for `Pipeline filtered out by workflow rules` error.
@@ -250,7 +253,8 @@ remotes table.
 - Setting private Gitlab token
 - Error message on linting error
 
-[Unreleased]: https://github.com/Blarc/gitlab-template-lint-plugin/compare/v1.11.1...HEAD
+[Unreleased]: https://github.com/Blarc/gitlab-template-lint-plugin/compare/v1.12.0...HEAD
+[1.12.0]: https://github.com/Blarc/gitlab-template-lint-plugin/compare/v1.11.1...v1.12.0
 [1.11.1]: https://github.com/Blarc/gitlab-template-lint-plugin/compare/v1.11.0...v1.11.1
 [1.11.0]: https://github.com/Blarc/gitlab-template-lint-plugin/compare/v1.10.0...v1.11.0
 [1.10.0]: https://github.com/Blarc/gitlab-template-lint-plugin/compare/v1.9.0...v1.10.0